requirements.txt
aiogram~=2.14
environs~=9.5.0
pandas
asyncpg~=0.27.0
Pillow~=9.5.0